[stats] StreamIds should be ints, not longs
Summary:
This not only makes counters a bit more efficient, but
also allows for proper generation of key names such that
we can avoid seeing negative stream ID values.

Test Plan: Check counter names in MD.

Reviewers: ethan

Subscribers: eng-team-list

Differential Revision: https://phab.fatline.io/D4195

+++ src/org/jitsi/impl/neomedia/rtp/remotebitrateestimator/RemoteBitrateEstimatorSingleStream.java
@@ -245,7 +245,7 @@ public class RemoteBitrateEstimatorSingleStream
         } // synchronized (critSect)
     }
 
-    private void recordArrivalFilterStats(long ssrc, long deltaArrivalTimeMs, double deltaDepartureTimeMs,
+    private void recordArrivalFilterStats(int ssrc, long deltaArrivalTimeMs, double deltaDepartureTimeMs,
             long deltaSizeBytes, double estimatedGroupDelayVariationMs, int numOfDeltas)
     {
         double deltaGroupDelayVariationMs = deltaArrivalTimeMs - deltaDepartureTimeMs;
